backend/requirements.txt
flask
flask-cors
pdfkit
xhtml2pdf
mysql-connector-python
